requirements.txt
# ComfyUI Art 项目依赖
fastapi>=0.95.0
uvicorn>=0.22.0
redis>=4.5.4
pydantic>=2.0.0
docker>=6.1.0
websockets>=11.0.3
requests>=2.28.2
python-dotenv>=1.0.0
loguru>=0.7.0
PyYAML>=6.0
asyncio>=3.4.3
aiohttp>=3.8.4
ulid-py>=1.1.0
anyio~=4.4.0
typing_extensions~=4.11.0